+++ src/components/Field/FormFieldDropdownSingleSelection/FormFieldDropdownSingleSelection.tsx
@@ -9,7 +9,7 @@ import { MosaicFieldProps } from "@root/components/Field";
 import { CustomPopperProps, DropdownData, DropdownSingleSelectionInputSettings } from "./FormFieldDropdownSingleSelectionTypes";
 
 // Components
-import InputWrapper from "../../components/InputWrapper";
+import InputWrapper from "../../InputWrapper";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore"
 import TextField from "@mui/material/TextField";
 import { MosaicLabelValue } from "@root/types";
